+++ androidx-kt/src/main/java/me/panpf/androidxkt/app/Fragmentx.kt
@@ -107,7 +107,7 @@ inline fun FragmentActivity?.findFragmentByViewPagerCurrentItem(viewPagerCurrent
 inline fun android.support.v4.app.Fragment?.findFragmentByViewPagerCurrentItem(viewPagerCurrentItem: Int): android.support.v4.app.Fragment? = Fragmentx.findFragmentByViewPagerCurrentItem(this, viewPagerCurrentItem)
 
 
-fun android.app.Fragment.requireContext(): Context = Fragmentx.requireContext(this)
+inline fun android.app.Fragment.requireContext(): Context = Fragmentx.requireContext(this)
 
 
 inline fun android.support.v4.app.Fragment.requireAppContext(): Context = Fragmentx.requireAppContext(this)