+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * :zap: Romx support Flyme 7.0.1
 * :sparkles: Fragmentx adds requireContext, requireAppContent method
+* :sparkles: Activityx adds appContext method
 
 
 ## v0.3.2